A few more small refactorings.
